# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AImpactScanner API
  version: 1.0.0
  description: |
    Analyze websites for AI search visibility using the MASTERY-AI Framework v3.1.1.

    ## Authentication

    All endpoints require authentication via one of:
    - **Bearer token**: JWT obtained from Supabase Auth (`Authorization: Bearer <token>`)
    - **API key**: Passed in the `X-API-Key` header

    ## Rate Limits

    Rate limits are per-day and vary by subscription tier:

    | Tier   | Calls/day |
    |--------|-----------|
    | Free   | 10        |
    | Coffee | 100       |
    | Growth | 1,000     |
    | Scale  | 10,000    |

    Response headers on every request:
    - `X-RateLimit-Limit` -- daily quota for your tier
    - `X-RateLimit-Remaining` -- approximate remaining calls
    - `X-RateLimit-Reset` -- Unix timestamp when the limit resets

    ## Monthly Analysis Limits

    In addition to API rate limits, scan submissions (`POST /v1/scan`) are capped
    per calendar month:

    | Tier   | Scans/month |
    |--------|-------------|
    | Free   | 3           |
    | Coffee | 10          |
    | Growth | 40          |
    | Scale  | 100         |
  contact:
    name: AI Search Mastery
    url: https://aimpactscanner.com
servers:
  - url: https://aimpactscanner-backend-production.up.railway.app
    description: Production
  - url: http://localhost:3001
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/scan:
    post:
      operationId: submitScan
      tags: [Scan]
      summary: Submit a URL for AI visibility analysis
      description: |
        Fetches the page, runs all MASTERY-AI Framework v3.1.1 factors, and
        returns the complete scan result including scores, action items, schema
        analysis, and readability metrics. Typical processing time is 5-15 seconds.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                  example: https://example.com
      responses:
        "200":
          description: Scan completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      url:
                        type: string
                        format: uri
                      status:
                        type: string
                        enum: [complete]
                      overall_score:
                        type: integer
                        minimum: 0
                        maximum: 100
                      grade:
                        $ref: "#/components/schemas/Grade"
                      factors:
                        type: array
                        items:
                          $ref: "#/components/schemas/FactorResult"
                      pillars:
                        $ref: "#/components/schemas/PillarBreakdown"
                      action_items:
                        type: array
                        items:
                          $ref: "#/components/schemas/ActionItem"
                      schema_analysis:
                        $ref: "#/components/schemas/SchemaResult"
                      readability:
                        $ref: "#/components/schemas/ReadabilityResult"
                      processing_time_ms:
                        type: integer
        "400":
          description: Invalid or missing URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                missing:
                  value: { success: false, error: "URL is required", code: "INVALID_REQUEST" }
                invalid:
                  value: { success: false, error: "Invalid URL format. Please provide a valid HTTP or HTTPS URL.", code: "INVALID_URL" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Page could not be fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Monthly analysis limit reached
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    properties:
                      used:
                        type: integer
                      limit:
                        type: integer
                      tier:
                        type: string
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/scan/{id}:
    get:
      operationId: getScan
      tags: [Scan]
      summary: Retrieve scan results by ID
      description: Returns a previously completed scan. Only scans owned by the authenticated user are accessible.
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Scan found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      url:
                        type: string
                        format: uri
                      status:
                        type: string
                        enum: [processing, completed, failed]
                      overall_score:
                        type: integer
                        nullable: true
                      grade:
                        $ref: "#/components/schemas/Grade"
                      factors:
                        type: array
                        items:
                          $ref: "#/components/schemas/FactorResult"
                      pillars:
                        $ref: "#/components/schemas/PillarBreakdown"
                      created_at:
                        type: string
                        format: date-time
                      completed_at:
                        type: string
                        format: date-time
                        nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/scan/{id}/actions:
    get:
      operationId: getScanActions
      tags: [Scan]
      summary: Get action items for a scan
      description: |
        Returns action items and schema data for a scan. Note: action items are
        generated at scan time and included in the `POST /v1/scan` response.
        For historical scans, this endpoint may return empty arrays unless the
        items were persisted separately.
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Action items retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      scan_id:
                        type: string
                        format: uuid
                      actions:
                        type: array
                        items:
                          $ref: "#/components/schemas/ActionItem"
                      schema:
                        $ref: "#/components/schemas/SchemaResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # --------------------------------------------------------------------------
  # Schema endpoint
  # --------------------------------------------------------------------------
  /api/v1/schema/generate:
    post:
      operationId: generateSchema
      tags: [Schema]
      summary: Generate schema markup for a URL
      description: Fetches the page and analyzes existing structured data, identifies missing schemas, and generates JSON-LD recommendations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                  example: https://example.com
      responses:
        "200":
          description: Schema analysis completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      url:
                        type: string
                        format: uri
                      schema_analysis:
                        $ref: "#/components/schemas/SchemaResult"
        "400":
          description: Invalid or missing URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Page could not be fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

  # --------------------------------------------------------------------------
  # Monitor endpoints (placeholders)
  # --------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key issued from your dashboard
  parameters:
    ScanId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Scan identifier
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: Unauthorized
            code: UNAUTHORIZED
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: Not found
            code: NOT_FOUND
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: An unexpected error occurred
            code: INTERNAL_ERROR
  schemas:
...
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
          description: Machine-readable error code
    Grade:
      type: string
      enum: ["A+", A, "A-", "B+", B, "B-", "C+", C, "C-", "D+", D, F]

    # -- Scan result types ----------------------------------------------------
    FactorResult:
      type: object
      required:
        - factor_id
        - factor_name
        - pillar
        - score
        - confidence
        - weight
        - evidence
        - recommendations
      properties:
        factor_id:
          type: string
          example: schema_markup
        factor_name:
          type: string
          example: Schema Markup Quality
        pillar:
          type: string
          description: MASTERY-AI pillar code
          enum: [M, A, S, T, E, R, Y, AI, P, TS]
        phase:
          type: string
          enum: [instant, background]
        score:
          type: number
          minimum: 0
          maximum: 100
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        weight:
          type: number
          description: Factor weight within its pillar (0-1 range)
          example: 0.85
        evidence:
          type: array
          items:
            type: string
          example:
            - "Found 2 JSON-LD blocks"
            - "Organization schema detected"
        recommendations:
          type: array
          items:
            type: string
          example:
            - "Add FAQ schema for question-based content"
        severity:
          type: string
          enum: [blocker, warning, info, ok]
        processing_time_ms:
          type: integer
    PillarBreakdown:
      type: object
      description: Weighted scores grouped by MASTERY-AI pillar
      additionalProperties:
        type: object
        properties:
          score:
            type: integer
          factorCount:
            type: integer
      example:
        AI: { score: 72, factorCount: 3 }
        M: { score: 85, factorCount: 2 }
    ActionItem:
      type: object
      required:
        - id
        - factor_id
        - factor_name
        - what
        - why
        - impact
        - fix_type
        - fix
      properties:
        id:
          type: string
        factor_id:
          type: string
        factor_name:
          type: string
        what:
          type: string
          description: Brief description of the issue
        why:
          type: string
          description: Why this matters for AI visibility
        impact:
          type: string
          enum: [high, medium, low]
        fix_type:
          type: string
          enum: [code, content, config]
        fix:
          type: string
          description: Step-by-step fix instructions
        code_snippet:
          type: string
          description: Ready-to-use code when fix_type is "code"

    # -- Schema analysis types ------------------------------------------------
    SchemaResult:
      type: object
      required:
        - detected
        - missing
        - generated
        - validation
        - completeness_score
      properties:
        detected:
          type: array
          items:
            $ref: "#/components/schemas/DetectedSchema"
        missing:
          type: array
          items:
            $ref: "#/components/schemas/MissingSchema"
        generated:
          type: array
          items:
            $ref: "#/components/schemas/GeneratedSchema"
        validation:
          type: array
          items:
            $ref: "#/components/schemas/SchemaValidation"
        completeness_score:
          type: number
          minimum: 0
          maximum: 100
...
    ReadabilityResult:
      type: object
      nullable: true
      properties:
        overall_score:
          type: number
          minimum: 0
          maximum: 100
        factors:
          type: array
          items:
            $ref: "#/components/schemas/ReadabilityFactor"
        top_improvements:
          type: array
          items:
            $ref: "#/components/schemas/ReadabilityImprovement"
...